% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCNVScopePublicData.R
\name{loadCNVScopePublicData}
\alias{loadCNVScopePublicData}
\title{Load CNVScope Neuroblastoma base data}
\usage{
loadCNVScopePublicData()
}
\value{
None. Data is loaded. Test using ls().
}
\description{
Loads the CNVScope data required to run the server, locally.
}
\examples{
loadCNVScopePublicData()
}
\keyword{genomic}
\keyword{matrix}
